+++ src/main/java/com/proyectoIntegrador/gameShare/repositorio/PlataformaRepositorio.java
@@ -5,5 +5,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface PlataformaRepositorio extends JpaRepository<Plataforma, Long> {
+public interface PlataformaRepositorio extends JpaRepository<Plataforma, Integer> {
 }
